Reworked the camera and updated the player
The camera now utilizes a FollowTarget script, so that it's no longer necessary to have the same follow code in all of the different states.
Added CrippledState and HurtState, in which the player will be slower. The player reaches these when they fall a substantial distance. They wear off after time, bringing the player back to the WalkingState.
Made the player's "boots" more slippery, so that they don't get stuck to colliders