Basic picture taking
Can now take a picture with the PictureCamera, which is forced to a 1-1 aspect ratio (looking incorrect in the game view, but correct on the image)

Known issues:
Cannot look up and down with the Picture Camera yet.
Cannot save multiple pictures yet (this implementation will be coupled to the scrapbooks picture implementation)
Camera doesn't properly adjust the game view.