Resolved layout group anchor issues
Layout group forces anchor to upper corner, creating issues for the pictures in the scrapbook.
Reworked the system with a custom override for the pictures.
Can now place pictures from the panel and remove them from the book, bringing them back to the panel.
Can now remove text from the scrapbook