Test feedback fixes - 1
Can now close scrapbook and camera with Escape
Photos can now be discarded
An icon/button makes it much easier to find and control the photo collection panel
Pages now have numbers. On the first page, the previous button disappears and on the last page, the next button.
Can now walk while in camera mode and crouch, but not sprint or jump.
Setup for camera flash animation (not properly working yet, but doesn't break the game)