UI fixes
Text is now on it's own layer (the lowest layer) and is immediately placed on this layer. All photos and stickers are above all text.
Bug where photo wasn't returned to the correct spot in the scrapbook when dragging it from the scrapbook on an occupied quest interaction interface resolved.